Check a11y tree when background pdf tab regains focus.
When a tab is unfocused, we won't be abled to receive accessibility
events which means we would wait for that event forever even when the pdf
accessibility tree is fully loaded. In order to amend that, we will do a
manually a11y tree scan when the tab is focused only if
is_pdf_a11y_info_loaded_ is false. A11y tree might not be ready when the
tab is getting focus so we will retry 3 seconds later at most 5 times
for each OnWebContentsFocused.